+++ .config/Dockerfile
@@ -83,6 +83,10 @@ ENV VATSIM_KAFKA_BROKER=${VATSIM_KAFKA_BROKER}
 ENV VATSIM_KAFKA_USER=${VATSIM_KAFKA_USER}
 ENV VATSIM_KAFKA_PASSWORD=${VATSIM_KAFKA_PASSWORD}
 ENV VATSIM_KAFKA_GROUP=${VATSIM_KAFKA_GROUP}
+ENV NAVIGRAPH_CLIENT_ID=${NAVIGRAPH_CLIENT_ID}
+ENV NAVIGRAPH_CLIENT_SECRET=${NAVIGRAPH_CLIENT_SECRET}
+ENV NAVIGRAPH_SERVER_ID=${NAVIGRAPH_SERVER_ID}
+ENV NAVIGRAPH_SERVER_SECRET=${NAVIGRAPH_SERVER_SECRET}
 
 COPY package.json package.json
 COPY yarn.lock yarn.lock
