+++ src/utils/backend/navigraph.ts
@@ -9,6 +9,7 @@ import { fromServerLonLat } from '~/utils/backend/vatsim';
 import { handleH3Exception } from '~/utils/backend/h3';
 import type { FullUser } from '~/utils/backend/user';
 import type { NavigraphGate, NavigraphRunway } from '~/types/data/navigraph';
+import { $fetch } from 'ofetch';
 
 function base64URLEncode(str: Buffer) {
     return str
@@ -161,6 +162,8 @@ export async function getNavigraphGates({ user, icao, event }: {
     event?: H3Event;
 }) {
     let gates: NavigraphGate[] = [];
+
+    if (!navigraphCurrentDb || !navigraphCurrentDb) return gates;
 
     try {
         await new Promise<void>((resolve, reject) => {
